+++ src/com/intellij/idea/plugin/hybris/system/bean/psi/BSPatterns.kt
@@ -20,6 +20,7 @@ package com.intellij.idea.plugin.hybris.system.bean.psi
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.system.bean.model.Bean
+import com.intellij.idea.plugin.hybris.system.bean.model.Property
 import com.intellij.patterns.PlatformPatterns
 import com.intellij.patterns.StandardPatterns
 import com.intellij.patterns.XmlAttributeValuePattern
@@ -58,4 +59,25 @@ object BSPatterns {
                 .withLocalName("extends")
         )
         .inside(beansXmlFile)
+
+    val BEAN_CLASS: XmlAttributeValuePattern = XmlPatterns.xmlAttributeValue(Bean.CLASS)
+        .withSuperParent(
+            2,
+            XmlPatterns.xmlTag()
+                .withLocalName("bean")
+        )
+        .inside(beansXmlFile)
+
+    val BEAN_PROPERTY_TYPE: XmlAttributeValuePattern = XmlPatterns.xmlAttributeValue(Property.TYPE)
+        .withSuperParent(
+            2,
+            XmlPatterns.xmlTag()
+                .withLocalName("property")
+                .withParent(
+                    XmlPatterns.xmlTag()
+                        .withLocalName("bean")
+                )
+        )
+        .inside(beansXmlFile)
+
 }